+++ jetpack_vendor/automattic/jetpack-forms/CHANGELOG.md
@@ -21,6 +21,7 @@ This is an alpha version! The changes listed here are not final.
 - Forms: elevate outline style notched labels with z-index so they are more easily selectable
 - Forms: labels in animated style support html
 - Forms: Update integration card to TypeScript.
+- Forms: Update single modal cards to TypeScript.
 
 ## [1.3.0] - 2025-06-16
 ### Added